+++ core/fxge/fx_font.h
@@ -79,7 +79,7 @@ FX_RECT GetGlyphsBBox(const std::vector<TextGlyphPos>& glyphs,
                       FontAntiAliasingMode anti_alias);
 
 ByteString GetNameFromTT(pdfium::span<const uint8_t> name_table, uint32_t name);
-size_t GetTTCIndex(pdfium::span<const uint8_t> font_data, size_t font_offset);
+uint32_t GetTTCIndex(pdfium::span<const uint8_t> font_data, size_t font_offset);
 
 inline bool FontStyleIsForceBold(uint32_t style) {
   return !!(style & pdfium::kFontStyleForceBold);
